backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ZakPorat"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.31"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
psycopg2 = "^2.9.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
argon2-cffi = "^23.1.0"
pip = "^24.1.1"
install = "^1.3.5"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
